Fix broken <span> at beginning of highlighted diff
If a regex matches on lines that begin with a space, it will create a
<span> tag directly at teh beginning of the code. The return statment
removes the first char which in this case will be "<" so the html tag
will be broken and there will be a left over space.

If we don't add the space in the first place, we don't have to remove
it.

This looks like some crazy workaround, so it might cause problems to
appear.

Signed-off-by: Florian Pritz <bluewind@xinu.at>