Allow to keep and reclaim uploads without being logged in
If a user keeps the browser open until his session expires and then
tries to upload something we now add it to the database, add the ID to
the new session and when someone logs in with that session the ID is
assigned. Until then even if you guess it correctly, you won't be able
to download it.

If the user still manages to let the 2nd session expire because he can't
find his password, the upload will be lost. Shit happens.

Signed-off-by: Florian Pritz <bluewind@xinu.at>