Rework file deletion; allow to delete multiple IDs at once
This removes the old form which was used to delete a single upload and
replaces it with checkboxes on the history page. All checked IDs will be
removed at once, instead of requiring the user to click through multiple
pages.

The old file/delete page is kept for compatibility with CLI clients.

Signed-off-by: Florian Pritz <bluewind@xinu.at>