Move user_logged_in check to controller constructor
TL;DR: Allows us to show a proper error page if encryption_key is
missing from the config.

muser->logged_in() can load the session class which will die if
encryption_key is not set in the config causing an error to be
displayed.

Because the header is also loaded when we display an error
loading the class will be tried again. CI maintains an array with
information which classes have been tried to be loaded and will simply
return true without loading again.

muser->logged_in() will then try to access $this->session which doesn't
exist. Since all of this happens when we are already in the header the
error message appears in the navigation being hard to read.

Signed-off-by: Florian Pritz <bluewind@xinu.at>