Fix handling of php errors
When the exception handler moved to a dedicated class
_exception_handler was no longer defined so CI defined it with it's own
code which does not throw an exception but rather just prints the less
helpful CI error page.

This could be triggered by deleting an upload while the cache_backend
is set to "file" and the cache does not hold the filedata for the
upload being deleted. In that case PHP throws a warning (unlink on a
missing file) which would no longer be caught with try/catch in the
cache backend's delete() function. Now the error is catchable again.

Signed-off-by: Florian Pritz <bluewind@xinu.at>