Fix redirect URI when using multiple tabs
If we store only the last called URI in the session we can't support
multiple browser tabs that all need to log in again. Fix this by
storing the URI in the URL.

Also change a trim() to ltrim() so that the URI string we store keeps
it's trailing slash.

Signed-off-by: Florian Pritz <bluewind@xinu.at>