API 2.0: Remove private fields from file/history
Since this is a breaking change bump the api version to 2.

The private fields are user_id and multipaste_id which where leaked via
the multipaste_items field. This commit also adds a test case to both
api versions that checks the returned fields.

NOTE: Most of this commit is copied from the files of api v1 so when
viewing the diff use --find-copies-harder for an easy to read diff.

Signed-off-by: Florian Pritz <bluewind@xinu.at>