Fix handling of %10 in paste content
CI tries to remove invisible escape chars, but this breaks handling of
code like the following:

if (m == 2 && (y%4 == 0 && y%100 != 0)|| y%400 == 0)

When pasted via the client all is fine, but when pasted in the webui CI
strips '%10' from the string and leaves 'y0 != 0'.

The second parameter of remove_invisible_characters indicates whether
the string is urlencoded so I believe that setting it to false should
be fine. This only disables the code that removes % escapes.

Signed-off-by: Florian Pritz <bluewind@xinu.at>