Fix redirect on login after registering new account
Previously the login box in the navigation would redirect to the
current page, but this page will throw an error in the case of the
registration page since that's the page with the invition key and that
key is no longer valid.

Fix this by redirecting to the $redirect_uri and ensure that this value
is set for all requests.

Signed-off-by: Florian Pritz <bluewind@xinu.at>