Fix double linebreaks when copying code in firefox
Apparently firefox doesn't like html code in <pre> and adds incorrect
linebreaks when copying the code. This does not happen with chromium.

Fix it by switching to <code> instead of <pre>. Also change the CSS to
refer to the element by class instead of by type so future changes are
simpler.

Not sure why, but in chroium changing the container element type breaks
the ordering of the lines. Adding display: table to the .code-container
fixes this.

References:
http://stackoverflow.com/questions/24851173/why-does-text-copied-from-a-set-of-pre-tags-have-double-line-breaks#comment41552816_26408656

Signed-off-by: Florian Pritz <bluewind@xinu.at>