Fix missing empty lines in chromium when copying code
Chromium apparently requires a <br> or some text to be present for a
line to be copied to the clipboard.

Firefox interprets the \n at the end as line break and copies that if
there is no content. Adding an explicit <br> makes chromium work
correctly, but firefox still converts the \n. Fix this by putting the
\n in an HTML comment. That way the output is still somewhat readable,
but both browsers do the correct thing.

Signed-off-by: Florian Pritz <bluewind@xinu.at>