history: Disable lazy loading for the first 42 thumbnails
This speeds up the initial rendering of the page. For a long history
(> ~1000 items) the lazyloading JS creates a noticeable delay (>500ms)
depending on the cpu power of the machine.

Disable lazy loading for the first items so that these are loaded
normally by the browser parallel to the scripts execution.

Signed-off-by: Florian Pritz <bluewind@xinu.at>