Fix broken lineendings when rendering ascii colors
Without -m ansi2html may not close the span tags at the end of the
line. Since we wrap the output lines in our own tags those left-open
tags break the rendering. Fix it by letting ansi2html wrap lines into
tags on it's own. This adds superflous tags in the HTML, but it works
nicely.

Signed-off-by: Florian Pritz <bluewind@xinu.at>