Stop making PDF thumbnails with imagemagick
Arch Linux disables PDF support in imagemagick due to security concerns.
This results in broken thumbnails for PDF files. By disabling it we just
get the normal file list which should be fine too. If necessary this
could be extended to be configurable, but I don't think doing that is
necessary.

Signed-off-by: Florian Pritz <bluewind@xinu.at>