Disable CSRF token regeneration for multipaste AJAX
If regeneration is on, the token is recreated after the first AJAX
submit and subsequent ajax submits or normal form submits break. By
disabling it here, we limit potential security issues to only this page,
but it also only works if the user does not submit any other forms while
they are on the AJAX page.

Signed-off-by: Florian Pritz <bluewind@xinu.at>