Pygments: Skip lexers that do not have names
With pygments 2.4.0 at least VBScript doesn't have a name and thus
causes an exception when the first name should be used. We use the name
to tell pygmentize which lexer to use so listing a lexer that doesn't
have a name doesn't work and thus there is no point in showing it in the
list.

Signed-off-by: Florian Pritz <bluewind@xinu.at>