thumbnail-view.js: Detect automatic image rotation in recent browsers
The default value has been changed to `from-image` and at least with
Firefox 76 and Chromium 83 this leads to double rotated images.
Disabling the automatic rotation with `image-orientation: none` in CSS
leads to incorrectly scaled images, (likely, but unverified) because the
width/height are returned incorrectly/rotated in jquery/javascript.
Whatever it may be, the easier fix is to just check for the new default
value and if the browser handles orientation automatically, we disable
our own code.

https://github.com/w3c/csswg-drafts/issues/3799

Signed-off-by: Florian Pritz <bluewind@xinu.at>