docker: Always regenerated config files on container startup
We should really regenerate the configs, run migrations and hooks during
each container startup. Otherwise you really can't update anything at
all.

The only thing that we do not do multiple times is adding the admin
user, but the add_user method already checks for that and simply
displayed an error. Hiding that error is probably not worth the effort
so we just add a quick explanatory comment before the potential error.

Signed-off-by: Florian Pritz <bluewind@xinu.at>