API: Only consider public methods as API endpoints
We are not only interested in checking if a method exists here, but
really also if it can be called (e.g. if it is public). Private methods
should not be considered as potential API endpoints. Before this,
private methods could be called, leading to a 500 error due to a call to
the private method.

Signed-off-by: Florian Pritz <bluewind@xinu.at>