fix(multipaste): Handle missing items in multipaste queue
This fixes an exception if the multipaste queue contains an item that
does not exist any more, e.g. because it has been deleted after it was
added to the queue.

`Exception 1/1 'ErrorException' with message 'Trying to access array offset on value of type bool' in /srv/http/filebin/application/controllers/file/Multipaste.php:103`

Signed-off-by: Florian Pritz <bluewind@xinu.at>