fix(duser_db): Guard against NULL values
When a user is deleted, their details are set to NULL. When no username
is specific for a password reset, this can trigger an email being sent
to an empty recipient which will not be deliverable.

Just to be safe, guard against NULL values for all the user related
functions.

Signed-off-by: Florian Pritz <bluewind@xinu.at>